refactor: move setSprinkler to its own file
