+++ src/game-logic/reducers/index.test.js
@@ -29,7 +29,6 @@ import {
   PURCHASEABLE_COMBINES,
   PURCHASEABLE_COW_PENS,
   SCARECROW_ITEM_ID,
-  SPRINKLER_ITEM_ID,
   STORAGE_EXPANSION_AMOUNT,
 } from '../../constants'
 import { huggingMachine, sampleCropItem1 } from '../../data/items'
@@ -1953,62 +1952,6 @@ describe('changeCowBreedingPenResident', () => {
             daysUntilBirth: COW_GESTATION_PERIOD_DAYS,
           },
         })
-      })
-    })
-  })
-})
-
-describe('setSprinkler', () => {
-  let state
-
-  beforeEach(() => {
-    state = {
-      field: [[null]],
-      fieldMode: fieldMode.SET_SPRINKLER,
-      inventory: [testItem({ id: 'sprinkler', quantity: 1 })],
-      itemsSold: {},
-      selectedItemId: SPRINKLER_ITEM_ID,
-    }
-  })
-
-  describe('plot is not empty', () => {
-    test('does nothing', () => {
-      const inputState = { ...state, field: [[testCrop()]] }
-      state = fn.setSprinkler(inputState, 0, 0)
-      expect(state).toEqual(inputState)
-    })
-  })
-
-  describe('plot is empty', () => {
-    test('sets sprinkler', () => {
-      const { field, inventory } = fn.setSprinkler(state, 0, 0)
-
-      expect(field[0][0]).toEqual(getPlotContentFromItemId('sprinkler'))
-      expect(inventory).toHaveLength(0)
-    })
-
-    describe('multiple sprinkler units remaining', () => {
-      test('updates state', () => {
-        const { fieldMode: newFieldMode, selectedItemId } = fn.setSprinkler(
-          { ...state, inventory: [testItem({ id: 'sprinkler', quantity: 2 })] },
-          0,
-          0
-        )
-        expect(newFieldMode).toBe(fieldMode.SET_SPRINKLER)
-        expect(selectedItemId).toBe(SPRINKLER_ITEM_ID)
-      })
-    })
-
-    describe('one sprinkler unit remaining', () => {
-      test('updates state', () => {
-        const { fieldMode: newFieldMode, selectedItemId } = fn.setSprinkler(
-          state,
-          0,
-          0
-        )
-
-        expect(newFieldMode).toBe(fieldMode.OBSERVE)
-        expect(selectedItemId).toBe('')
       })
     })
   })
